Essential fixes to make CHDatastructures build for MINGW64 using gcc.
1) Disable unknown pragma warnings.
2) Link with GNUstep CoreBase.
3) Change casts to unsigned long to uintptr_t as __WIN64 (aka __LLP64__) has different sizeof (unsigned long) and sizeof (void *).
4) Use rand_s() on Win32 instead of arc4random().

+++ source/CHAbstractBinarySearchTree.m
@@ -890,7 +890,7 @@ CHBinaryTreeNode* CHCreateBinaryTreeNodeWithObject(id anObject) {
 			{
 			NSLog (@"Current node %p. Finished object enumeration", current);
 			[(id) state -> extra [3] release];	/* Release the object enumerator */
-			state -> extra [3] = (unsigned long) nil;		/* Mark it deleted so we know we're not in object enumeration mode */
+			state -> extra [3] = (uintptr_t) nil;	/* Mark it deleted so we know we're not in object enumeration mode */
 			current = current -> right;		/* Advance to point to the next object, and continue through the tree */
 			}
 		}
